Make regulation groups always to belong to a plan
Regulation groups cannot belong to multiple plans so make it clear to which plan it belongs to.

General regulation groups use the same `plan_id` to define the relationship even though there is a little nuance difference in the meaning of "owner" plan and a plan "in which this regulation group is a general regulation group". To find general regulation groups for a plan one must filter its regulation groups by their type.
This change makes it possible for a plan to have multiple general regulation groups.